+++ examples/react/tanstack-start/src/utils/form.tsx
@@ -3,7 +3,7 @@ import {
   ServerValidateError,
   createServerValidate,
   getFormData,
-} from '@tanstack/react-form/start'
+} from '@tanstack/react-form-start'
 import { setResponseStatus } from '@tanstack/react-start/server'
 import { formOpts } from './form-isomorphic'
 
@@ -16,11 +16,8 @@ const serverValidate = createServerValidate({
   },
 })
 
-export const handleForm = createServerFn({
-  method: 'POST',
-  response: 'raw',
-})
-  .validator((data: unknown) => {
+export const handleForm = createServerFn({ method: 'POST' })
+  .inputValidator((data: unknown) => {
     if (!(data instanceof FormData)) {
       throw new Error('Invalid form data')
     }
@@ -30,19 +27,12 @@ export const handleForm = createServerFn({
     try {
       const validatedData = await serverValidate(ctx.data)
       console.log('validatedData', validatedData)
-      // Persist the form data to the database
-      // await sql`
-      //   INSERT INTO users (name, email, password)
-      //   VALUES (${validatedData.name}, ${validatedData.email}, ${validatedData.password})
-      // `
     } catch (e) {
       if (e instanceof ServerValidateError) {
         // Log form errors or do any other logic here
         return e.response
       }
 
-      // Some other error occurred when parsing the form
-      console.error(e)
       setResponseStatus(500)
       return 'There was an internal error'
     }